src/table: Make proper module by declaring header
It's custom in C++ to separate files as so-called "compilation units". One
unit is a C++-file, and all the files it directly includes. The compilation
unit is the unit that has to recompile if something changes, so for a big
project it's good to have many small compilation-units. Then a change in
one unit can be recompiled in isolation, without rebuilding everything. For
really big projects, compilation times can be 30 minutes+.